moved modified files in core
