+++ MTPHandler/__init__.py
@@ -1,4 +1,4 @@
-from .modified import Plate, Protein, Reactant
+from .core import Plate, Protein, Reactant
 
 __URL__ = ""
 __COMMIT__ = ""
