+++ MTPHandler/core/__init__.py
@@ -1,25 +1,15 @@
 from .plate import Plate
-from .well import Well
-from .initcondition import InitCondition
-from .vessel import Vessel
-from .abstractspecies import AbstractSpecies
 from .protein import Protein
 from .reactant import Reactant
-from .speciestype import SpeciesType
-from .sboterm import SBOTerm
-from .datatypes import DataTypes
+from .well import Well
+from .initcondition import InitCondition
 
 __doc__ = ""
 
 __all__ = [
     "Plate",
     "Well",
+    "Species",
     "InitCondition",
-    "Vessel",
-    "AbstractSpecies",
-    "Protein",
-    "Reactant",
     "SpeciesType",
-    "SBOTerm",
-    "DataTypes",
 ]
