--- MTPHandler/core/datatypes.py
@@ -1,10 +1,0 @@
-from enum import Enum
-
-
-class DataTypes(Enum):
-    CONCENTRATION = "conc"
-    ABSORPTION = "abs"
-    FEED = "feed"
-    BIOMASS = "biomass"
-    CONVERSION = "conversion"
-    PEAK_AREA = "peak-area"
