+++ MTPHandler/core/speciescondition.py
