+++ MTPHandler/ioutils/enzymeml.py
@@ -5,10 +5,10 @@ from datetime import datetime
 
 from CaliPytion import Standard
 
-from MTPHandler.modified.initcondition import InitCondition
-from MTPHandler.modified.reactant import Reactant
-from MTPHandler.modified.protein import Protein
-from MTPHandler.modified.well import Well
+from MTPHandler.core.initcondition import InitCondition
+from MTPHandler.core.reactant import Reactant
+from MTPHandler.core.protein import Protein
+from MTPHandler.core.well import Well
 
 # Specify EnzymeML version
 URL = "https://github.com/EnzymeML/enzymeml-specifications.git"
