--- MTPHandler/modified/__init__.py
@@ -1,15 +1,0 @@
-from .plate import Plate
-from .protein import Protein
-from .reactant import Reactant
-from .well import Well
-from .initcondition import InitCondition
-
-__doc__ = ""
-
-__all__ = [
-    "Plate",
-    "Well",
-    "Species",
-    "InitCondition",
-    "SpeciesType",
-]
