--- MTPHandler/modified/speciestype.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class SpeciesType(Enum):
-    BUFFER = "buffer"
-    ENZYME = "enzyme"
-    SUBSTRATE = "substrate"
-    PRODUCT = "product"
-    INHIBITOR = "inhibitor"
